QuFin Moduelar V.5/backend/Backend/requirements.txt
pandas==2.0.3
beautifulsoup4==4.11.2
black==23.1.0
Django==3.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
httpx==0.24.1
isort==5.12.0
openpyxl==3.0.10
orm-choices==1.0.0
pyexcel==0.7.0
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
pyflakes==3.0.1
python-dateutil==2.8.2
selenium==4.7.2
xlrd==2.0.1
xlwt==1.3.0
psycopg2-binary==2.9.5
celery==5.2.7
redis==4.5.4
pydantic==1.10.7
django-cors-headers==3.14.0
pytest==7.3.1
pytest-django==4.5.2
python-dotenv==1.0.1